4-1 DivByZero 버그 수정

+++ app/src/main/java/com/example/project4_1/MainActivity.kt
@@ -81,7 +81,7 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun doCalc(stat: String) {
-        if (edit2.text.toString().toInt() == 0) {
+        if (edit2.text.toString().toDouble() == 0.0) {
             Toast.makeText(this, R.string.div_zero, Toast.LENGTH_SHORT).show()
             return
         }
